playgameコントローラを分離
