--- src/main/resources/templates/gameroom/playgame.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" xmlns:sec="http://www.springframework.org/schema/security">
-
-<head>
-  <meta charset="utf-8">
-  <title>Quiz_7/ゲームルーム選択</title>
-  <link rel="stylesheet" href="/css/origin.css" />
-</head>
-
-<body>
-  <div class="container">
-    <div class="display">
-      <h3>公開ゲームルーム</h3>
-      <div class="scroll-full">
-        <table th:if="${publicGameroomList}">
-          <tr>
-            <th>ゲームルーム名</th>
-            <th>ホスト名</th>
-            <th>最大人数</th>
-            <th>リンク</th>
-          </tr>
-          <tr th:each="pgameroom : ${publicGameroomList}">
-            <td th:text="${pgameroom.gameRoomName}"></td>
-            <td th:text="${pgameroom.hostUserName}"></td>
-            <td align="right" th:text="${pgameroom.maxPlayers}"></td>
-            <td><a th:href="@{/playgame/join_check(room=${pgameroom.gameRoomID})}">選択</a></td>
-          </tr>
-        </table>
-      </div>
-    </div>
-
-    <div class="links-section">
-      <a href="../main">戻る</a>
-    </div>
-  </div>
-</body>
-
-</html>
