+++ src/main/resources/templates/playgame.html
@@ -4,31 +4,33 @@
 <head>
   <meta charset="utf-8">
   <title>Quiz_7/ゲームルーム選択</title>
-  <link rel="stylesheet" href="/css/origin.css">
+  <link rel="stylesheet" href="/css/origin.css" />
 </head>
 
 <body>
   <div class="container">
     <div class="display">
       <h3>公開ゲームルーム</h3>
-      <table th:if="${publicGameroomList}">
-        <tr th:each="gameroom : ${publicGameroomList}">
-          <td th:text="${gameroom.roomName}"></td>
-          <td th:text="${gameroom.hostName}"></td>
-          <td><a th:href="@{/playgame/join_check(room=${gameroom.ID})}">選択</a></td>
-        </tr>
-      </table>
-
-      <form th:action="@{/gameroom/prepare_open(room=${gameroom.ID})}" method="post" id="max_players"
-        class="form-inline">
-        <label for="max_players">最大参加人数： </label>
-        <input type="number" name="max_players" size="5" value="50" min="1" max="50" style="width:80px;" required />
-      </form>
+      <div class="scroll-full">
+        <table th:if="${publicGameroomList}">
+          <tr>
+            <th>ゲームルーム名</th>
+            <th>ホスト名</th>
+            <th>最大人数</th>
+            <th>リンク</th>
+          </tr>
+          <tr th:each="pgameroom : ${publicGameroomList}">
+            <td th:text="${pgameroom.gameRoomName}"></td>
+            <td th:text="${pgameroom.hostUserName}"></td>
+            <td align="right" th:text="${pgameroom.maxPlayers}"></td>
+            <td><a th:href="@{/playgame/join_check(room=${pgameroom.gameRoomID})}">選択</a></td>
+          </tr>
+        </table>
+      </div>
     </div>
 
-    <div class="input-form">
-      <button type="submit" form="max_players" class="colored_button">公開</button><br />
-      <a href="../gameroom" class="button">戻る</a>
+    <div class="links-section">
+      <a href="../main">戻る</a>
     </div>
   </div>
 </body>
